+++ src/Resources/ResourceManager/SdkExtensions/ResourcesExtensions.cs
@@ -256,7 +256,8 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkExtensions
 
                 foreach (KeyValuePair<string, DeploymentVariable> pair in dictionary)
                 {
-                    result.AppendFormat(rowFormat, pair.Key, pair.Value.Type, pair.Value.Value.ToString().Indent(maxNameLength + maxTypeLength + 4).Trim());
+                    result.AppendFormat(rowFormat, pair.Key, pair.Value.Type, 
+                        JsonConvert.SerializeObject(pair.Value.Value).Indent(maxNameLength + maxTypeLength + 4).Trim());
                 }
             }
 
@@ -317,12 +318,26 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkExtensions
                 if (properties.Outputs != null)
                 {
                     Dictionary<string, DeploymentVariable> outputs = JsonConvert.DeserializeObject<Dictionary<string, DeploymentVariable>>(properties.Outputs.ToString());
+                    // Continue deserialize if the type of Value in DeploymentVariable is array
+                    outputs?.Values.ForEach(dv => {
+                        if ("Array".Equals(dv?.Type))
+                        {
+                            dv.Value = JsonConvert.DeserializeObject<object[]>(dv.Value.ToString());
+                        }
+                    });
                     deploymentObject.Outputs = outputs;
                 }
 
                 if (properties.Parameters != null)
                 {
                     Dictionary<string, DeploymentVariable> parameters = JsonConvert.DeserializeObject<Dictionary<string, DeploymentVariable>>(properties.Parameters.ToString());
+                    // Continue deserialize if the type of Value in DeploymentVariable is array
+                    parameters?.Values.ForEach(dv => {
+                        if ("Array".Equals(dv?.Type))
+                        {
+                            dv.Value = JsonConvert.DeserializeObject<object[]>(dv.Value.ToString());
+                        }
+                    });
                     deploymentObject.Parameters = parameters;
                 }